added campaign + list folder to separate components
