+++ src/components/campaign/CampaignCreateForm.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import loginGate from './requires-login-gate';
+import loginGate from '../requires-login-gate';
 import CampaignCreateSelect from './CampaignCreateSelect';
 
 export class CampaignCreateForm extends Component {
