+++ src/components/campaign/CampaignCreateSelect.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import Select from 'react-select';
 
-import loginGate from './requires-login-gate';
+import loginGate from '../requires-login-gate';
 
 import 'react-select/dist/react-select.css';
