+++ src/components/campaign/CampaignItem.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 
-import { deleteCampaign } from '../actions/userActions';
+import { deleteCampaign } from '../../actions/userActions';
 
 export function CampaignItem(props) {
