+++ src/components/campaign/Dashboard_Campaigns.js
@@ -2,10 +2,10 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 
-import loginGate from './requires-login-gate';
+import loginGate from '../requires-login-gate';
 import CampaignItem from './CampaignItem.js';
 
-import { getCampaigns } from '../actions/userActions';
+import { getCampaigns } from '../../actions/userActions';
 
 export class DashboardCampaigns extends Component {
