+++ src/components/list/Contact_CreateForm.js
@@ -1,14 +1,14 @@
 import React, { Component } from 'react';
 import {Field, reduxForm, focus} from 'redux-form';
 
-import Input from './Input';
+import Input from '../Input';
 
-import { required, nonEmpty } from '../validators';
+import { required, nonEmpty } from '../../validators';
 
-export class ListCreateForm extends Component {
+export class Contact_CreateForm extends Component {
 
     onSubmit(values) {
-        this.props.createNewList(values);
+        this.props.createContact(values);
     }
 
     render() {
@@ -32,16 +32,32 @@ export class ListCreateForm extends Component {
 
         return(
             <section>
-                <form className='login-form' onSubmit={this.props.handleSubmit( values => this.onSubmit(values))}>
+                <form className='login-form' onSubmit={this.props.handleSubmit(values => this.onSubmit(values))}>
 
                     {error}
                     {message}
-                    <label htmlFor="name">Name</label>
+                    <label htmlFor="name">First Name:</label>
                     <Field
                         component={Input}
                         type="text"
-                        name="name"
-                        id="name"
+                        name="firstName"
+                        id="firstName"
+                        validate={[required, nonEmpty]}
+                    />
+                    <label htmlFor="name">Last Name:</label>
+                    <Field
+                        component={Input}
+                        type="text"
+                        name="lastName"
+                        id="lastName"
+                        validate={[required, nonEmpty]}
+                    />
+                    <label htmlFor="name">Email:</label>
+                    <Field
+                        component={Input}
+                        type="text"
+                        name="email"
+                        id="email"
                         validate={[required, nonEmpty]}
                     />
                     <button disabled={this.props.pristine || this.props.submitting}>
@@ -56,4 +72,4 @@ export class ListCreateForm extends Component {
 export default reduxForm({
     form: 'listCreateForm',
     onSubmitFail: (errors, dispatch) => dispatch(focus('login', 'username'))
-})(ListCreateForm);
+})(Contact_CreateForm);
