+++ src/components/list/Dashboard_List_Create.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 
-import loginGate from './requires-login-gate';
+import loginGate from '../requires-login-gate';
 import ListCreateForm from './ListCreateForm';
 
-import { createList } from '../actions/userActions';
+import { createList } from '../../actions/userActions';
 
 export class DashboardListCreate extends Component {
     constructor(props){
