+++ src/components/list/Dashboard_Lists.js
@@ -2,10 +2,10 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 
-import loginGate from './requires-login-gate';
+import loginGate from '../requires-login-gate';
 import ListItem from './ListItem.js';
 
-import { getLists } from '../actions/userActions';
+import { getLists } from '../../actions/userActions';
 
 export class DashboardLists extends Component {
