+++ src/components/list/ListCreateForm.js
@@ -1,14 +1,14 @@
 import React, { Component } from 'react';
 import {Field, reduxForm, focus} from 'redux-form';
 
-import Input from './Input';
+import Input from '../Input';
 
-import { required, nonEmpty } from '../validators';
+import { required, nonEmpty } from '../../validators';
 
-export class Contact_CreateForm extends Component {
+export class ListCreateForm extends Component {
 
     onSubmit(values) {
-        this.props.createContact(values);
+        this.props.createNewList(values);
     }
 
     render() {
@@ -32,32 +32,16 @@ export class Contact_CreateForm extends Component {
 
         return(
             <section>
-                <form className='login-form' onSubmit={this.props.handleSubmit(values => this.onSubmit(values))}>
+                <form className='login-form' onSubmit={this.props.handleSubmit( values => this.onSubmit(values))}>
 
                     {error}
                     {message}
-                    <label htmlFor="name">First Name:</label>
+                    <label htmlFor="name">Name</label>
                     <Field
                         component={Input}
                         type="text"
-                        name="firstName"
-                        id="firstName"
-                        validate={[required, nonEmpty]}
-                    />
-                    <label htmlFor="name">Last Name:</label>
-                    <Field
-                        component={Input}
-                        type="text"
-                        name="lastName"
-                        id="lastName"
-                        validate={[required, nonEmpty]}
-                    />
-                    <label htmlFor="name">Email:</label>
-                    <Field
-                        component={Input}
-                        type="text"
-                        name="email"
-                        id="email"
+                        name="name"
+                        id="name"
                         validate={[required, nonEmpty]}
                     />
                     <button disabled={this.props.pristine || this.props.submitting}>
@@ -72,4 +56,4 @@ export class Contact_CreateForm extends Component {
 export default reduxForm({
     form: 'listCreateForm',
     onSubmitFail: (errors, dispatch) => dispatch(focus('login', 'username'))
-})(Contact_CreateForm);
+})(ListCreateForm);
